+++ src/app/admin-module/components/recipe-list.component.ts
@@ -50,7 +50,7 @@ import { DataCollectionService } from '../services/data-collection.service';
               <td class="px-6 py-4">
                 <a
                   [routerLink]="['/', 'admin', 'page', item.id]"
-                  class="text-white bg-gray-800 hover:bg-gray-900 focus:outline-none focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-gray-800 dark:hover:bg-gray-700 dark:focus:ring-gray-700 dark:border-gray-700"
+                  class="text-white bg-gray-800 hover:bg-gray-900 focus:outline-none focus:ring-4 focus:ring-gray-300 font-medium rounded text-sm px-4 py-2 mr-2 mb-2 dark:bg-gray-800 dark:hover:bg-gray-700 dark:focus:ring-gray-700 dark:border-gray-700"
                   >Detail</a
                 >
               </td>